Updated mediawiki/extensions Project: mediawiki/extensions/Buggy  bd4a76c3fcddbbde299b2e935dd84ee2accf9bec
Make JSHint pass

Use undef:false to tolerate undefined references.

It's still useful to have JSHint here so that it is known
what errors are intentionally allowed. As well as to ensure
general syntax compliance aside from code quality.

If and when invalid syntax will be tested as well, just add
the filepath to a jshintignore file.

Change-Id: Ia8c9817107e08389db338254be37529cb0744a10