Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  1960dad342664fa44a1d6e4698f9b6ebbbca3eca
Introduce wikibase.entityIdFormatter

This wraps entity fetching and calling the EntityId formatting functions from
`wikibase.utilities.ui` in new service classes and passes them to the views.

It is a necessary prerequisite for being able to swap out the EntityId
formatting logic with a different implementation. It also frees the views from
knowing how to format EntityIds and removes the duplicate EntityId formatting
code.

Change-Id: I4a2fdfbc3faa8a640683006111539812ab764faf