Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  85eabcf6ddf908931ebf428047f5becf96f05469
Fix filtering in change dispatcher.

This fixes a bug that causes changes to be dispatched to wikis
to which it is not relevant. The issue was that the return value of
isRelevantChange(), which returns a boolean, was checked with !== null,
which of course is always true, even if the return value was false.

NOTE: If5ea19e29a15 refactors changeDispatcher and adds tests.

Change-Id: Ib6f733fbd84e27784c99a2d1440a35f625215d2b