Updated mediawiki/extensions Project: mediawiki/extensions/Translate  00bd24b9170e72d1a6369a37d772303d2c9db5e2
Position the group selector to appear at the bottom of the trigger.

The group selector hide the trigger when opened due to its height growth
at the bottom when the data is loaded. Added collision:none so that
group selector is opened to the bottom side of the trigger and height
change would not cause a problem.

Change-Id: I6e08fcb635ddc8170383d91927d1a2d6eb4be8bf