Updated mediawiki/extensions Project: mediawiki/extensions/Flow  fa9220f74c6f3677f4671a4460b7b5c8c2feddcb
Provide ConfirmEdit with the previous revision content

A recent change to the ConfirmEdit extension is failing
our unit tests.  The simplest way around is to explicitly
provide the previous revision content along with the current
content.  This was likely already a bug that we never noticed
previously, ConfirmEdit did not have access to this previous
content.

Add more information to WikitextException thrown by
Flow\Parsoid\Utils->parser.

Requires Ifc93cbaa in ConfirmEdit to be merged as well (merged).

Bug: T93956
Change-Id: If508cab599ac069be17ada99c0c582bced6adfdc