Updated mediawiki/extensions Project: mediawiki/extensions/Flow  60fd8c9926a82e4958db57c95a076852eb587576
Store data for IRCLine tests

The ->newNextRevision calls in that dataprovider end up at
AbstractRevision::getContent, which will try to fetch the
revision's collection & title (in order to parse content
for the correct title)

It will try to locate these things in storage, but they
weren't there.

The reason we never had this was that content was already
stored in html, so this was bypassed.

Bug: T94025
Change-Id: I713c16fe5d2cfb482ac01acec5d922a7902ca61d