Updated mediawiki/extensions Project: mediawiki/extensions/Flow  60d76d9229a5d84f50870e6d2f2325101205d426
First store workflow, then the rest

ReferenceRecorder (on of the post listeners) will need content
in HTML, for which it will need to retrieve Collection & Title,
to figure out what page to parse against.

Since workflow hasn't yet been stored, this fails.

Obvious solution here is to first store workflow. However, the
comment explicitly said it was there for a reason.
I'm not entirely certain that reason is still valid. I tried
creating new topics & stuff, still seemed to work just fine.
I'm hoping whoever reviews this, remembers exactly why this was!

If we can't merge this, we could remove referenceloader from
post listeners if storage format is not html.

Bug: T94025
Change-Id: Ic82a64207dd92f5461fd9651ace4b444478095d6