Updated mediawiki/extensions Project: mediawiki/extensions/Flow  57c875b941f05508934168dfe2d728d4d11cd2f3
Don't expect cuc_id to exist

There's 3 possible paths in CheckUser that could lead to this hook
being triggered. 1 of those doesn't provide cuc_id.

Funny thing: the only thing we need cuc_id for, is to log good info
when something went wrong. The only thing that went wrong, was exactly
this...

Bug: T90285
Change-Id: Ifab3379727f347b40f25025fc0601e3e7e54e39e