Updated mediawiki/extensions Project: mediawiki/extensions/OpenStackManager  997664d26bc8986a82c5df9184920c5d0b6efaa3
Don't reset default puppet config when editing puppet config.

Because, hey, maybe I want to /change/ the default settings.  Also,
this resulted in a failured due to a duplicate setting even if
the web UI left a default the same as it was.

The defaults are set at creation time in OpenStackNovaHost.  That's
really the only place they should be set.

Change-Id: Ib81a87c1e28be805628c96e83961e7b2104b8892