Updated mediawiki/extensions Project: mediawiki/extensions/Flow  e2b8ae5ace72e78284c25b283bac595f38fa6c73
Add content type: fixed-html

API is drawing a lot on view-* APIs, both for displaying content
& editing content (we're now directly using the html output in VE)

We need to be able to request 2 kinds of HTML: the one including
our ContentFixer adjustments (for output) & the one for VE, where
we need the original content.

This patch adds a 'fixed-html' content type. 'html' will from now
be the original, unfixed, html - the way VE needs it & we store it.
'fixed-html' is what we'll output, where ContentFixers are applied.

Bug: T90681
Change-Id: I3506c60790b03bf368a24d0964a81bda593c6aaf