Updated mediawiki/extensions Project: mediawiki/extensions/DonationInterface  33e7c5959af3a1a402bcb94416c1d3cf30b46da1
Update required version of Monolog to 1.12.0

To stay in sync with version in core/vendor.
Corresponding change in DonationInterface/vendor is Ibe85af85460cc85
That and the follow-on change to clean up unlintable files should
be merged and the submodule updated in the deployment branch during
the next DI deploy.

Change-Id: I5a5f9e916a381e65f251c503433cd04d413ac9eb