Updated mediawiki/extensions Project: mediawiki/extensions/InviteSignup  17601094e975768d09f2071214549d479fe49ed8
Fix bug: when $wgEmailConfirmToEdit is true, account creation would
fail with a "No email address" error.

This is due to a check in SpecialUserlogin.php, described as follows:
"if you need a confirmed email address to edit, then obviously you
need an email address."

With this configuration, an account creation form without an email
address is not allowed. The only straightforward workaround I can see
is to disable $wgEmailConfirmToEdit on the Special:Userlogin page when
setting $wgInviteSignupHash.

This should keep the configuration and InviteSignup working at the same
time.

Change-Id: I21479b19b8f2f2a009bb41ca01b76533400576cb