Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  7ef1b8152e4005b6bbc2de7b5bee040590e37ad3
Get rid of $wgMFPageSections config

This variable was used on two places to determine if we are in
mobile mode. Both places can use more intelligent conditions.

To place the toc, simply looking up if a #toc already exists and
replacing it is much easier.

For toggling of sections no check is needed at all because on the
skin's desktop version, there is a div under #content wrapping
the whole content so that the selector >h1,>h2,>h3,>h4,>h5,>h6
wont work in that place anyways and therefore the condition is not
needed.

Bug: T95884
Change-Id: Icc586d9d76f81f512e2a94d326a7f28168bc44ef