Updated mediawiki/extensions Project: mediawiki/extensions/BlueSpiceExtensions  f72a7e62ba03bc182f4eb92ee22a3269df34efc1
Fixed image crispness on resize

When an image was resized, the saved size differed in several pixels from
the displayed size. The reason was a padding added to the image element in
the editor. This is now removed, together with the margin, which is also
not rendered by the wiki.

Change-Id: Ied701b7ddadb029b646d4a47df7b2daba83481cb
(cherry picked from commit bf50b248508283e4e2ab579bbd8acf5d0fe4378a)