Updated mediawiki/extensions Project: mediawiki/extensions/BlueSpiceExtensions  e4ab5716f7d541ab24c1f6af3aae7db745f11343
Fixed image crispness on resize

When an image was resized, the saved size differed in several pixels from
the displayed size. The reason was a padding added to the image element in
the editor. This is now removed, together with the margin, which is also
not rendered by the wiki.

Change-Id: Ied701b7ddadb029b646d4a47df7b2daba83481cb