Updated mediawiki/extensions Project: mediawiki/extensions/Flow  6bd7352c5f9a52003cfa18c8627df28700416b6a
Special:EnableFlow archives pre-existing talk page

Enabling flow on an existing talk page is now accepted.
It tries to archive the talk page to a subpage.

The name of the archive is configurable using a new
form field. The default value is %s/Archive_%d and
it is localized.

It uses a template named "Wikitext talk page converted to Flow"
to add a reference to the archive page at the top of
the flow board.

It uses a template named "Archive for converted wikitext talk page"
to add a reference to the flow board at the top of
the archive page.

Bug: T72073
Change-Id: If3b23383e4289290cde1237b74a4bbbb6bdcb455