Updated mediawiki/extensions Project: mediawiki/extensions/SpamDiffTool  b8d1166c6af80b58e85ca9d62de87b478f6f47c0
Fix fatal when diffs are being viewed on a special page

Calling DifferenceEngine in the context of a SpecialPage invokes the
DiffViewHeader hook and thus -- somewhat correctly -- the "add to spam"
link is also generated. But clicking on the link causes a PHP fatal,
because the Title object returned by $diffEngine->getTitle() points to the
special page which invoked DifferenceEngine, not to the non-special page
that was actually edited.

ShoutWiki SVN r3074

Change-Id: I464723d5cb59bf324d7916969b5a0e33a7d631d5