Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  9ef81381bec8013a0422c48d6fe6b37329d5ea84
Convert browser tests that check for create page link

Converts a variety of tests that check for the 'create a new page' link
to exist/not exist to use the api.  These no longer check for that link,
but its tested extensively in the remaining browser tests and shouldn't
be anything to be concerned about.

Change-Id: I0be02c8a1b55e51c96d389a3ba1cbb145412eed8