Updated mediawiki/extensions Project: mediawiki/extensions/Babel  12faaf20e38e3de660e6fe7038b2a7376ada5274
Explicitly define module position

Style modules currently added through addModuleStyles default
to being in the head ("top" position). This is an unhealthy default,
since only critical styles that are needed at pageload should be
in the head. In order to be able to switch the default to "bottom",
existing module positions have to be defined explicitly.

Bug: T97410
Change-Id: I43bd351ce2a31d722907a8e652e4b74ff7532a31