Updated mediawiki/extensions Project: mediawiki/extensions/Calendar  e02b955f886925e9eb6a62cfbb1d7a69611dacd0
Explicitly define module position

Style modules currently added through addModuleStyles default
to being in the head ("top" position). This is an unhealthy default,
since only critical styles that are needed at pageload should be
in the head. In order to be able to switch the default to "bottom",
existing module positions have to be defined explicitly.

Bug: T97410
Change-Id: I092a8153d6653485ba9ecfc362c614c0587b14e0