Updated mediawiki/extensions Project: mediawiki/extensions/CodeReview  4d03a1106ceed40077ed1dcde8c36f5e645d7613
Explicitly define module position

Style modules currently added through addModuleStyles default
to being in the head ("top" position). This is an unhealthy default,
since only critical styles that are needed at pageload should be
in the head. In order to be able to switch the default to "bottom",
existing module positions have to be defined explicitly.

Bug: T97410
Change-Id: Iefe4eb5ce41c2880d4b1a59305129855fa6038d1