Updated mediawiki/extensions Project: mediawiki/extensions/GettingStarted  dce20921888a027f9ed6eb17217c0533ea629bb4
Explicitly define module position

Style modules currently added through addModuleStyles default
to being in the head ("top" position). This is an unhealthy default,
since only critical styles that are needed at pageload should be
in the head. In order to be able to switch the default to "bottom",
existing module positions have to be defined explicitly.

Bug: T97410
Change-Id: I4558219988116345dc2efc9180d9c8295f92ca68