Updated mediawiki/extensions Project: mediawiki/extensions/GlobalUserPage  8f949e5db759f2d75f734d0e68f6fc0be38033c9
Explicitly define module position

Style modules currently added through addModuleStyles default
to being in the head ("top" position). This is an unhealthy default,
since only critical styles that are needed at pageload should be
in the head. In order to be able to switch the default to "bottom",
existing module positions have to be defined explicitly.

Bug: T97410
Change-Id: Ib15bf92078c9294322b94482716c14f9321a8932