Updated mediawiki/extensions Project: mediawiki/extensions/MassMessage  ca3244ade5e90f6634ee90227008361fdb5c9135
Explicitly define module position

Style modules currently added through addModuleStyles default
to being in the head ("top" position). This is an unhealthy default,
since only critical styles that are needed at pageload should be
in the head. In order to be able to switch the default to "bottom",
existing module positions have to be defined explicitly.

Bug: T97410
Change-Id: Id23cd8415c93c3ed88623508cc217ca68e195653