Updated mediawiki/extensions Project: mediawiki/extensions/UploadWizard  9d0c456b23a35b10d947df0088ec0922f0fc9755
Explicitly define module position

Style modules currently added through addModuleStyles default
to being in the head ("top" position). This is an unhealthy default,
since only critical styles that are needed at pageload should be
in the head. In order to be able to switch the default to "bottom",
existing module positions have to be defined explicitly.

Bug: T97410
Change-Id: I0d47205fdf724a58af754ef10403771e4c46d853