Updated mediawiki/extensions Project: mediawiki/extensions/ZeroBanner  7ad1be9e94deb7938c950491851128e67f536e5b
Explicitly define module position

Style modules currently added through addModuleStyles default
to being in the head ("top" position). This is an unhealthy default,
since only critical styles that are needed at pageload should be
in the head. In order to be able to switch the default to "bottom",
existing module positions have to be defined explicitly.

Bug: T97410
Change-Id: If7fe170901992b4a302ff70d43ac0a7f4a185580