Updated mediawiki/extensions Project: mediawiki/extensions/WikimediaIncubator  8f33ba8b966ad6918e716fac2707dcc602a000b1
Explicitly define module position

Style modules currently added through addModuleStyles default
to being in the head ("top" position). This is an unhealthy default,
since only critical styles that are needed at pageload should be
in the head. In order to be able to switch the default to "bottom",
existing module positions have to be defined explicitly.

Bug: T97410
Change-Id: Iff7e58b095c083ec40a7c9c73ebf3c3bc403305c