Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  7ee1f948a60fc0f8b8ce855294b1110d99f3ea72
Upload files via api in cucumber tests

Bypasses the browser and uploads files directly via api. Trys to be
smart and checks sha1 hash of existing file before uploading a new
version. Not entirely thrilled with this code, but rubocop really
really wanted 10 line methods, so this is what i came up with.

Change-Id: I0ab5e8d9e9d6988d7534a845b4510e49a92f71fc