Updated mediawiki/extensions Project: mediawiki/extensions/BlueSpiceFoundation  1825fd26371d3ec4ac84a796bf3c67410bb20df4
API base classes: ExtJS store and Tasks

This change contains some refactoring of class names and also changes
internal implemenation details! This introduces a potential breaking
change for subclasses. But as those classes are not used very much
this should not be a problem.

The goal is to provide a common API infrastructure within BlueSpice.

* Also added a CSS fix for CRUDGridPanels when a grouping feature is being
used
* Added more dataset processing logic to ExtJSStore base class
* Created abstract 'makeData' and implemented 'execute' to have even less
code in the subclass
* Added some generic hooks
* Added standardized taskData for Tasks API

Change-Id: Id8e1ad8bd2a14e9f8bddbc5c6873b1522d294b31