Updated mediawiki/extensions Project: mediawiki/extensions/PhpTags  47466ede801610345cd75919d45ea7edf8a92123
Fix PhpTagsException::isFatal(), some refactoring (v5.0.1)

PhpTagsException::isFatal() checked $this->code, but for HookException,
$this->code was only set to the actual error level in the __toString()
method. The check was always false for HookException instances.

To solve this, HookException now overrides how PhpTagsException works,
and reuses a new formatMessage() method which was broken out in
PhpTagsException. The constant EXCEPTION_FROM_HOOK is no longer used,
and was removed.

A new isCatchable() method was also added to both exception classes,
and isFatal() now returns true for EXCEPTION_CATCHABLE_FATAL as well.
The Runtime still uses the same check for isFatal() instead of
!isCatchable(). I think this makes the most sense for the Runtime to
do, since an uncaught catchable fatal error is really a fatal error.
(It would be up to an extension to catch it before it reaches the
Runtime.)

Bug: T95323
Change-Id: If1b5d6734522bb0f6a289f10473b95e73e3ffc80