Updated mediawiki/extensions Project: mediawiki/extensions/SecurePoll  9596f517ab64edc64ca937297138492d93adcb70
Fixup script for bv2015 incorrect date range

populateEditCount.php used a date range which was too short. So, extend
the date range by scanning only the missing timestamp range, and adding
any resulting edits to the previous edit count.

Modify voterList.php to generate a new list, so that it can be run again
while the election is in progress.

Bug: T97924
Change-Id: Iae874c0403a8362929362ca645f4aca18feb0269