Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  6b2a5f63c2cbcb417fd569d7a891e6a9665af242
Null is a valid value in Searcher::$namespaces

The namespaces variable is nullable, set by SearchEngine::replacePrefixes().
This indicates that all namespaces should be searched.  Icd762939 introduced
type hints to CirrusSearch\Searcher::__construct which do not accept this
nullable value and causes exceptions. CirrusSearch appopriatly handles the null
case so mark it as such.

The replacement of false with array() looks to be safe, nothing checks for an
exact false boolean, only that the namespaces value itself is falsey. The value
is defaulted to null because thats the only way to tell php that null is an
allowed value when typehinted to array.

Bug: T98082
Change-Id: Ic0d6adcde305d9ed4ccd1edb0406080f616a393e