Updated mediawiki/extensions Project: mediawiki/extensions/UniversalLanguageSelector  d04ade35741aded3515bb2eb1057f3e6e3ef5499
Explicitly define module position

Style modules currently added through addModuleStyles default
to being in the head ("top" position). This is an unhealthy default,
since only critical styles that are needed at pageload should be
in the head. In order to be able to switch the default to "bottom",
existing module positions have to be defined explicitly.

Bug: T97410
Change-Id: I209f93a49ddd0aa143348c0238849a121856dce1