Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  b3a43cc293bb623937e3d8df1069db1bdd4b04a4
Fix errors that look like OR OR OR OR

These are caused by us pulling phrase prefix queries into a, well, phrase
prefix query so they don't get to participate in OR and AND. They didn't
work before so its not _that_ bad that they don't work now. But they shouldn't
throw an error.

T94814

Change-Id: I5a77bc602469eef4c3595f448c22ce4d542da13d