Updated mediawiki/extensions Project: mediawiki/extensions/DonationInterface  9ed7d3d7c440e49da66e72b29252c249e5926c47
Add Astropay to form settings for BR/CC

FIXME: This is a dummy entry to allow FormChooser to route to
Astropay.  When we have One True Form, the chooser should route
to the appropriate gateway based on gateway settings that look
like these country and currency lists, and form settings should
disappear (or be a simple whitelist of ffnames).

Bug: T90505
Change-Id: Ia3a985d531bf7b4836194eb167c4825a9c0e6f60