Updated mediawiki/extensions Project: mediawiki/extensions/Flow  e3b683f9e4c3c15ecd21ef6d59d7b418abdcaff2
Don't log exception when missing permissions, just ignore it

formatApi() fails on a couple of occasions, one of which being when
a user has insufficient permissions. That's an "acceptable" error:
we shouldn't log it, just ignore that row.
Some places already counter this by first checking permissions, then
passing it off to formatApi() - if that one fails, it's not a
permission issue & we should log the failure.
Let's just throw an exception right away if it's a real error case,
and return false if it's a "this is no error but we can't show you
the data" case. Makes dealing with this simpler.

One potential regression: this formatApi function is not just called
from RC, Contribs, ... formatters, but also in other places
(everywhere...), so adding a new exception in there might break
those other places.
The first check was already being logged if it was being hit, and I
couldn't find it in the logs; so pretty sure we won't see it happen.

Change-Id: Ibc45d1a81e023cd0942374fdc5e4fc6a978f936f