Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  d6c5b7e8b31ac3f863c067ba083ecea293ade1a1
Show a better error message, when the user can't edit a page due to the false group

If a page is protected and a user isn't allowed to edit this page, show a CtaDrawer,
instead of a toast with the message, that the page is protected to prevent vandalism and
give the user a cancel button and a button that redirects him to the history page to
get more information (e.g. the reason why the page was protected and to which group).

Bug: T95305
Change-Id: I59122b568cd353ca5c3aba0a28522effb15cfab4