Updated mediawiki/extensions Project: mediawiki/extensions/WikibaseJavaScriptApi  76452ebbd22bdb32b354c43ee523c4663ed98d1b
Update to data-values/javascript 0.7.0

Note: I'm not sure if the added dependency in the resources.php file is
actually needed. But on the other hand, if that's not needed, why is the
component even mentioned in the composer.json file? I think it should be
in both files or in none.

Bug: T99231
Change-Id: Ife23d87ac400ce4469d0f4997ca65818c493276a