Updated mediawiki/extensions Project: mediawiki/extensions/ContentTranslation  f2722152b53028352350c4e1d0867aee0efff128
Fix the directionality of the columns' sub-headings

This is UI and not content, so the general direction of this element
must match the user interface language.
The language label's direction must match the language's,
and its alignment must match the general UI,
so text-align is set to 'left' and it will be flipped.

Change-Id: Ifad2e9586acb5f090b55c3a6d27412c8b3588309