Updated mediawiki/extensions Project: mediawiki/extensions/Cargo  4dbf7dc126909b522536969fa1b7ef3f2278ac94
Renamed FULL_PRECISION and TIME_MISSING date precision constants

They are now DATE_AND_TIME and DATE_ONLY, respectively. Also, a field
of type 'Date' that has a full date is of precision DATE_ONLY (it used
to be FULL_PRECISION), which makes the correct display in the
'calendar' format easier to do.

Change-Id: Ib330d17cb22625c7c3116afb077a79ca2e7f8a8f