Updated mediawiki/extensions Project: mediawiki/extensions/examples  11fdea21d8d101fb410f2d8aeb9fbeb2bca9d6a6
add an addHelpLink() example

This commit adds an example of the new addHelpLink() functionality
introduced in MediaWiki 1.25. Because this is merely an example
extension, there is no actual help page for the special page in
question: as such we use the "How to become a MediaWiki hacker" page
as a placeholder.

Bug: T97619
Change-Id: I4d105a6b852359b3606f2a2f0077cfe2b604b65d