Updated mediawiki/extensions Project: mediawiki/extensions/WikidataPageBanner  d3ce650547d9a70baa5323d41ef30d52bf5179d8
WikidataPageBanner::getWikidataBanner added

A function getWikidataBanner() is added, which fetches a banner for an article
page if the page has an associated item page and a banner statement in wikidata.
The function returns the banner image name on wikimedia commons or 'null' if it
fails to achieve the above objective.

Bug: T98044
Change-Id: Ic05f6c132e4e5f4d96f0d986d068941eb27480a8