Updated mediawiki/extensions Project: mediawiki/extensions/SyntaxHighlight_GeSHi  58c81408f5da6a3e91d9875b62a47d0eeca543d4
Explicitly define module position

Style modules currently added through addModuleStyles default
to being in the head ("top" position). This is an unhealthy default,
since only critical styles that are needed at pageload should be
in the head. In order to be able to switch the default to "bottom",
existing module positions have to be defined explicitly.

Bug: T97410
Change-Id: Ie120a781ac1950abd7963d6f722aa316b5542b51