Updated mediawiki/extensions Project: mediawiki/extensions/Translate  21a9201ecb02daa1ddec35c8f070bacc5d3bacc8
Change default to UnsupportedTranslationAid

This ensures that new translation aids get correctly noted as unsupported
in all message groups (until they are marked as supported). For most
message groups, this will simply be a case of adding them to the central
list (TranslationAid::getTypes).

Also, make showAssistantLanguages() resiliant to errors.

Change-Id: I7bae04ee2e6cf27b21cf95b10f98bc3463a7648d