Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  5547fcaad0293066e3620a59d49438cea9a3b3c0
Enable CirrusSearch-PerUser pool counter in fully featured config

Long standing todo, the required patch has been merged some time ago.
The only question is if 1 is really the right limit here, for jenkins
it probably was but for a fully featured wiki configuration probably
not. As such it has been updated to a limit of 5.

Change-Id: I4ecf748e1ab28b78bd409aecb32bee81cd1b4047