Updated mediawiki/extensions Project: mediawiki/extensions/Gather  2072bb728dd461603b08a5741971bac006f55e67
Show related pages when editing 1 to 3 items collections

When editing a collection if it has one item it will show related pages to that
item to help creators of new collections fill their collections up.

Depends on having the CirrusSearch extension installed
* http://www.mediawiki.org/wiki/Extension:CirrusSearch
* vagrant roles enable cirrussearch

After those two extensions are installed, try out:
https://en.wikipedia.org/wiki/Special:ApiSandbox#action=query&prop=pageimages&format=json&piprop=thumbnail&pithumbsize=75&pilimit=3&generator=search&gsrsearch=morelike%3ABanana&gsrnamespace=0&gsrlimit=3

And change Banana with the name of an article on your wiki until you get some
results.

Then create a collection with that Article, and you should see related results.

Bug: T100634
Change-Id: I53c82175b66465177d057c3a012631f526f63df2