Updated mediawiki/extensions Project: mediawiki/extensions/Citoid  c8ca892de0beb6365618c298081082c902b3dd4a
Check templateField type & add QUnit tests

Previously, didn't check templateField type for type
Array when citoidField was also an array and would
attempt to add arrays as a template parameter.
This would cause VE to throw errors and the inspector
would be stuck on pending. Now, if there is a type
mismatch, it ignores the field and no longer tries to
add it as a template parameter.

Also adds QUnit tests - one added for 'good' template
data, including all three supported types (String,
Array, 2D array), 6 for mismatched/'bad' template
data- an event that could easily happen, as these
are not validated- and one for Objects in both the
citation data and templateData, neither of which
should occur in the wild.

Run tests using mediawiki's QUnit by going to:
Special:JavaScriptTest/qunit?module=ext.citoid

Bug: T100651
Change-Id: I0b897cace77ce47d952f4b53007404b13ea312d9