Updated mediawiki/extensions Project: mediawiki/extensions/VectorBeta  cbd610331b594d17f2756a7dc277bd391ba14890
Explicitly define module position

Style modules currently added through addModuleStyles default
to being in the head ("top" position). This is an unhealthy default,
since only critical styles that are needed at pageload should be
in the head. In order to be able to switch the default to "bottom",
existing module positions have to be defined explicitly.

Bug: T97410
Change-Id: I952c0f97c90ba7e0b931f6798dc5d1c8770614a5