Updated mediawiki/extensions Project: mediawiki/extensions/WikimediaEvents  8991081bd0416b689be525ba006bf371e846ddf8
Report time to first edit as `ttfe`

The time that elapses between a user creating an account and that same user
making her first edit is a good indicator of how good a job we are doing at
being encouraging, welcoming and comprehensible to new users.

One common web performance metric is "TTFB", or time-to-first-byte. This, then,
is "TTFE", or time-to-first-edit.

Bug: T99060
Change-Id: Ib61cd546ab8e70f252360ad22e6befecc38c6236