Updated mediawiki/extensions Project: mediawiki/extensions/CodeReview  2205a02a53d8595ae0fa2336699741597c5c5b0f
Explicitly define module position

Style modules currently added through addModuleStyles default
to being in the head ("top" position). This is an unhealthy default,
since only critical styles that are needed at pageload should be
in the head. In order to be able to switch the default to "bottom",
existing module positions have to be defined explicitly.

Bug: T97410
Change-Id: I90f6d31c9b478d34ddfc04d8ac0d5c34e7bad36f