Updated mediawiki/extensions Project: mediawiki/extensions/TimedMediaHandler  2ba344709451457912abcfa279e368348c9dc23a
Consistently use "mw", "mediaWiki" and "jQuery" as globals instead of properties

In most files this was the case already, but a few random files
used them as properties as window for no good reason.

This makes errors hard to debug because unlike the ReferenceError that
would throw an exception, if they are accessed as properties the 'undefined'
value is yielded and the error doesn't happen until the closure tries
to access a property of the local variable holding that undefined value.

Change-Id: I82fbba34aa2a326066895283b4d1f444317aa362