Updated mediawiki/extensions Project: mediawiki/extensions/Translate  dc9d45c9043e86f88a29c21509b3546f17ea54a4
Add afterEditorShown hook

This allows extensions to hook in very late in the display process,
once the DOM has been fully constructed and the editor is fully visible.
This ensures that e.g. all heights and widths are known. This fills a
similar role to that which afterRegisterFeatures did previously.

Change-Id: I549c60a5e684a309bcd4146aa7c953d9b6e050d1