Updated mediawiki/extensions Project: mediawiki/extensions/RSS  6421f1f37b1bcf9cfb956ed1cfe095c8a365a1e9
Don't rely on strip marker uniqueness

The RSS extension expects two different parser instances to use two different
string patterns for strip markers as a workaround for T36763. As of I31d4556bb,
the parser uses a fixed string pattern for strip markers, so this trick no
longer works. We can work around this by doing the same thing we did in
SemanticForms in I7ed3996a1, which is to perform the stripping / unstripping in
extension code directly.

Bug: T10104
Change-Id: Ice44740fb8fea9db4e29e8c2b411384e24a88ae3