Updated mediawiki/extensions Project: mediawiki/extensions/CentralAuth  eef53de16df78ba91328a116d4277c069bda99ed
Show log entries of previous renames on the GlobalRenameQueue

A certain user is constantly requesting renames, so it has been
requested to display previous renames on the queue view page. This will
display any renames to the requesting user's current username. If the
user has been in a chain of renames (A→B→C→D), only ones to the current
name will be shown.

Bug: T101080
Change-Id: I52ab8f7b9c712fec1fc1555b7d2e1b4ec222fcb1