Updated mediawiki/extensions Project: mediawiki/extensions/Calendar  bdd2ca6f6562fdd8725e9714ff19d46c923262bf
Explicitly define module position

Style modules currently added through addModuleStyles default
to being in the head ("top" position). This is an unhealthy default,
since only critical styles that are needed at pageload should be
in the head. In order to be able to switch the default to "bottom",
existing module positions have to be defined explicitly.

Bug: T97410
Change-Id: I4dba8e6ab372513c7686cce1323ff9356a9f07d7