Updated mediawiki/extensions Project: mediawiki/extensions/ContentTranslation  87b5261216d9e817febef053b186fb88b34327af
Show login page if the URL has campaign parameter for anon users

For non-logged-in users show login page instead of nosuchspecialpage
if the URL has the campaign parameter.
After logging in the user should be taken to Special:CX
again with all the query parameters.

This is important for campaigns like email campaigns, to prevent
a situation where the user who clicked is not be logged in
and doesn't see anything relevant.
Showing nosuchspecialpage is a bad idea in that case.

Bug: T101489
Change-Id: I600af89c91e14e45e58dcb17fa98d3cdda09122b