Updated mediawiki/extensions Project: mediawiki/extensions/ContentTranslation  9dbb5e5b87f5d23ca240137f5eea10e6859608a3
Callout widget: Improve the hover behavior

Avoid showing multiple callouts if the trigger is 'hover'.
Also hide the callout, if the mouse enters siblings of triggers,
if present.

For contribution menu, this improves the user experience.
Before this patch, moving the mouse over the contribution list
to the logout link horizontally did not cause the contribution
menu to hide. With this patch it will hide.

For the stats page, we can improve the custom hover handler because
the widget will make sure only one callout is shown. While fixing
that, fixed a small bug that the menu is coming only when tail for
bars (bar with ...) is present.

Bug: T99076
Change-Id: Iadadf68023eaa7aa8506e9076d80847eef64296c