Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  a03ce2b0393c38e3af15987e83882f617ef7a593
Fix EntityFactory calling a non-existing newEmpty method

This is required to make the DataModel 3 switch possible. We removed
Property::newEmpty but no tool had a chance to find this line of code.
This patch does not solve the issues in this class, but also does not
make it worse in my opinion.

Note that this class is deprecated anyway. If you think I'm making it
worse, please only block this patch if you have a better suggestion.

This patch is cross-compatible with all DataModel versions.

Change-Id: Ibd64d52c0900ea08835f6fa38aea87ebfc8276d7