Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  748608c04bd1d0f95f3c4ec5c070f47792935513
Make sitelinkgroupview.css more compatible with cached HTML

In I64d91e2 we changed the DOM. This is always a problem because the
HTML (which is cached for each individual page) sticks much longer to
all the caching layers involved than the CSS and JavaScript (which is
only cached once because it's shared across all pages). What usually
happens is that users see outdated HTML with new CSS and JavaScript.
This becomes much, much smoother if we keep the CSS and JavaScript
backwards compatible for a month.

Bug: T95889
Change-Id: I33f20dd11dfe5ece65cd176f1c3aa15bf7531889