Updated mediawiki/extensions Project: mediawiki/extensions/Flow  ef8c99d9ce24180e1999a594621382b61baed4aa
Improve Workflow::isNew check

We no longer need the hacky isNew property to accomplish what
isNew() does - we can now use the pageId now that it's guaranteed
to be associated with a workflow when it's stored.

Also removed misleading phpdoc. When saving to storage,
fromStorageRow is called (which sets isNew to false) to update
the data in cache (& memory).

Change-Id: I7a163d8a93cfb4a8f6e8e897722fab73c0e92624