Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  6003e00319ddfe7c44f381bd59a2e8e647012c2a
Fix selection of snak values (to not select invalid values)

Don't select and save snak value when it has become invalid.

Without this check, statementview is proceeding to save with
a stale value, even after the value became invalid (e.g. user
is continuing to select in entity selector).

As far as I can tell, this seems to be the problem in T98471.

Bug: T98471
Change-Id: If140ce69d5ab6bc52c8d7a9644490f5dabce75ff