Updated mediawiki/extensions Project: mediawiki/extensions/WikidataPageBanner  31d223bdfbab2823ddfef5380a0e878fb3454e80
WikidataPageBanner custom banner on page-preview

Modify loadModules() to add banner style after checking a valid namespace for
banner and not check for articlebanner property. This is being done because the
extension in any case tries to add a banner on allowed namespaces, and adding a
style assuming a banner, allows banners to be seen on page-preview.
Remove addition of 'loadImage' module as it no longer exists.

Bug: T100983
Change-Id: I3c69250a540d6ce1ad05f9e0bbe00c7db7069b2b